resolved conflicts, good for push to main work by Robert/Cherry
