+++ client/rawgApi.js
