+++ server/models/index.js
@@ -1,4 +1,6 @@
 const User = require("./User");
-const Thought = require("./Thought");
+const Comment = require("./Comment");
+const Game = require("./Game");
+const Review = require("./Review");
 
-module.exports = { User, Thought };
+module.exports = { User, Comment , Game , Review };
