+++ server/server.js
@@ -1,28 +1,38 @@
-require('dotenv').config({ path: '../.env'}); // load environment variables from .env file
-
-const express = require('express');
-const cors = require('cors');
-const mongoose = require('mongoose');
-
+const express = require("express");
+const { ApolloServer } = require("apollo-server-express");
+const path = require("path");
+// files in schemas folder
+const { typeDefs, resolvers } = require("./schemas");
+const db = require("./config/connection");
+const { authMiddleware } = require("./util/auth");
+const PORT = process.env.PORT || 3001;
 const app = express();
-const port = process.env.PORT || 4000;
-
-// connect to MongoDB
-mongoose.connect(process.env.MONGO_URI, {
-  useNewUrlParser: true,
-  useUnifiedTopology: true,
+const server = new ApolloServer({
+  typeDefs,
+  resolvers,
+  context: authMiddleware,
 });
-
-// middleware
-app.use(cors());
+app.use(express.urlencoded({ extended: false }));
 app.use(express.json());
-
-// routes
-app.get('/', (req, res) => {
-  res.send('Welcome to the RAWG API');
+if (process.env.NODE_ENV === "production") {
+  app.use(express.static(path.join(__dirname, "../client/build")));
+}
+app.get("/", (req, res) => {
+  res.sendFile(path.join(__dirname, "../client/"));
 });
-
-// start the server
-app.listen(port, () => {
-  console.log(`Server is listening on port ${port}`);
-});
+// Create a new instance of an Apollo server with the GraphQL schema
+const startApolloServer = async (typeDefs, resolvers) => {
+  await server.start();
+  server.applyMiddleware({ app });
+  db.once("open", () => {
+    app.listen(PORT, () => {
+      console.log(`API server running on port ${PORT}!`);
+      console.log(
+        `Use GraphQL at http://localhost:${PORT}${server.graphqlPath}`
+      );
+    });
+  });
+};
+// Call the async function to start the server
+// files in schema folder
+startApolloServer(typeDefs, resolvers);
